+++ FWLog.Data/Models/FreteTipo.cs
@@ -5,7 +5,7 @@ namespace FWLog.Data.Models
     public class FreteTipo
     {
         [Key]
-        public int IdFreteTipo { get; set; }
+        public long IdFreteTipo { get; set; }
         public string Descricao { get; set; }
         public string Sigla { get; set; }
     }
